Initial setup to have a mailman server running
This includes running containers for mailman-core and mailman-web, both
connecting to the same database used by the boost web site.

The branch proposes a minor refactoring for the shared usage of the
compose-start.sh script, allowing both the boost web server and the mailman-web
server to wait for the db, run migrations, collect assets and run the django
devserver.